Created a list of BankAccount's so that multiple accounts can be created and accessed via a password.
Whenever a new BankAccount is created, it is added to list, but my login method is iffy right now, Sometimes I am able to login to an account other times not. If you choose one if the admin options, it will print out the list with the users name, balance and password. The password to access the admin option is scsiz